refactor: helper 함수를 분리한다. - 각각 coin과 product로 분리
